move back to original file path
