+++ e2e/support/helpers/e2e-embedding-iframe-sdk-helpers.ts
@@ -1,13 +1,11 @@
 import type { MetabaseTheme } from "metabase/embedding-sdk/theme/MetabaseTheme";
 
 import { createApiKey } from "./api";
-import {
-  enableSamlAuth,
-  mockAuthProviderAndJwtSignIn,
-} from "./component-testing-sdk";
+import { mockAuthProviderAndJwtSignIn } from "./component-testing-sdk";
 import { setTokenFeatures } from "./e2e-enterprise-helpers";
 import { enableJwtAuth } from "./e2e-jwt-helpers";
 import { restore } from "./e2e-setup-helpers";
+import { enableSamlAuth } from "./embedding-sdk-testing";
 
 const EMBED_JS_PATH = "http://localhost:4000/app/embed.js";
